Unify seeders to garantee orderly insertion
